Do checks before rmtree

+++ select_model.py
@@ -920,7 +920,8 @@ def run_model_selection():
     else:
         model_name = dataset_name
     results_dir = '{}/{}'.format(args.out_dir, model_name)
-    rmtree(results_dir)
+    if os.path.exists(results_dir) and os.path.isdir(results_dir):
+        rmtree(results_dir)
     os.makedirs(results_dir, mode=0o755)
     # train w/ independent test sets
     if args.test_dataset:
